Flaten dependencies
Previously this Dockerfile was built on top of an exercism-base image
that was built on a debian-base image. This hierarchy made it difficult
to add/remove functionality. I resolved this be building the image from
the debian:9 image. Now all of the required functionality is
configured/installed when this image is built.